change spinner model

+++ pokemongui/src/main/java/aiki/gui/components/PaginatorEgg.java
@@ -191,7 +191,7 @@ public final class PaginatorEgg extends Paginator {
 
     @Override
     public void changeNbResults() {
-        int int_ = (Integer)getNbResults().getValue();
+        int int_ = getNbResults().getValue();
         if (int_ <= 0) {
             return;
         }
