+++ pokemongui/src/main/java/aiki/gui/components/PaginatorItem.java
@@ -294,7 +294,7 @@ public final class PaginatorItem extends Paginator {
 
     @Override
     public void changeNbResults() {
-        int int_ = (Integer)getNbResults().getValue();
+        int int_ = getNbResults().getValue();
         if (int_ <= 0) {
             return;
         }
