+++ pokemongui/src/main/java/aiki/gui/components/PaginatorMove.java
@@ -419,7 +419,7 @@ public final class PaginatorMove extends Paginator {
 
     @Override
     public void changeNbResults() {
-        int int_ = (Integer)getNbResults().getValue();
+        int int_ = getNbResults().getValue();
         if (int_ <= 0) {
             return;
         }
