+++ pokemongui/src/main/java/aiki/gui/components/PaginatorPokemon.java
@@ -437,7 +437,7 @@ public final class PaginatorPokemon extends Paginator {
 
     @Override
     public void changeNbResults() {
-        int int_ = (Integer)getNbResults().getValue();
+        int int_ = getNbResults().getValue();
         if (int_ <= 0) {
             return;
         }
